safari-534.53-branch/LayoutTests/fast/dom/anchor-getParameter.html

<!DOCTYPE html>
<html>
<body>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

var cases = [
    ["http://example.com/foo/bar?a=b", 'a'],
    ["http://example.com/foo/bar?abc=xy", 'abc'],
    ["http://example.com/foo/bar?a=x&b=y", 'a'],
    ["http://example.com/foo/bar?a=x&b=y", 'b'],
    ["http://example.com/foo/bar?a=x&a=y", 'a'],
    ["http://example.com/foo/bar?a=x&a=y&a=z", 'a'],
    ["http://example.com/foo/bar?a=x&A=y&a=z", 'A'],
    ["http://example.com/foo/bar?121=", '121'],
    ["http://example.com/foo/bar?a&", 'a'],
    ["http://example.com/foo/bar?a", 'a'],
    ["http://example.com/foo/bar?=a&", 'a'],
    ["http://example.com/foo/bar?a=b=c", 'a'],
    ["http://example.com/foo/bar?a=b=c", 'b'],
    ["http://example.com/foo/bar?a=b=&", 'a'],
    ["http://example.com/foo/bar?a=b=&", 'b'],
    ["http://example.com/foo/bar?&a=b", 'a'],
    ["http://example.com/foo/bar?&a=b&", 'a'],
    ["http://example.com/foo/bar?a=b#xyz", 'a'],
    ["http://example.com/foo/bar?#a=b#xyz", 'a']
];

document.write('<div style="display:none">');
for (var i = 0; i < cases.length; ++i)
    document.write('<a href="' + cases[i][0] + '">Link</a><br>');
document.write('</div>');

var elmts = document.getElementsByTagName('a');
for (var i = 0; i < elmts.length; ++i)
    document.write(elmts[i].href + ' (' + cases[i][1] + ') => ' + elmts[i].getParameter(cases[i][1]) + '<br>');
</script>
</body>
</html>
