LayoutTests/fast/dom/anchor-getParameter-expected.txt

http://example.com/foo/bar?a=b (a) => b
http://example.com/foo/bar?abc=xy (abc) => xy
http://example.com/foo/bar?a=x&b=y (a) => x
http://example.com/foo/bar?a=x&b=y (b) => y
http://example.com/foo/bar?a=x&a=y (a) => y
http://example.com/foo/bar?a=x&a=y&a=z (a) => z
http://example.com/foo/bar?a=x&A=y&a=z (A) => y
http://example.com/foo/bar?121= (121) => 
http://example.com/foo/bar?a& (a) => 
http://example.com/foo/bar?a (a) => 
http://example.com/foo/bar?=a& (a) => 
http://example.com/foo/bar?a=b=c (a) => b=c
http://example.com/foo/bar?a=b=c (b) => 
http://example.com/foo/bar?a=b=& (a) => b=
http://example.com/foo/bar?a=b=& (b) => 
http://example.com/foo/bar?&a=b (a) => b
http://example.com/foo/bar?&a=b& (a) => b
http://example.com/foo/bar?a=b#xyz (a) => b
http://example.com/foo/bar?#a=b%23xyz (a) => 
